+++ src/main/java/org/bytedeco/javacpp/FunctionPointer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2011-2013 Samuel Audet
+ * Copyright (C) 2011-2017 Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -25,6 +25,7 @@ package org.bytedeco.javacpp;
 import org.bytedeco.javacpp.annotation.ByPtr;
 import org.bytedeco.javacpp.annotation.ByRef;
 import org.bytedeco.javacpp.annotation.ByVal;
+import org.bytedeco.javacpp.annotation.ValueSetter;
 import org.bytedeco.javacpp.tools.Generator;
 
 /**
@@ -38,7 +39,13 @@ import org.bytedeco.javacpp.tools.Generator;
  * To use FunctionPointer, subclass and add a native method named {@code call()}
  * or {@code apply()}, along with its return type and parameters, as well as the
  * usual {@code native void allocate()} method to support explicit allocation,
- * which is typically a requirement for callback functions.
+ * which is typically a requirement for callback functions. We can implement a
+ * callback in Java by further subclassing and overriding {@code call/apply()}.
+ * <p>
+ * If you have an address to a native function, it is also possible to call it
+ * by defining a {@link ValueSetter} method with a single {@link Pointer} parameter,
+ * along with {@code native} declarations for {@code allocate()} and {@code call()}.
+ * After allocating the object and setting the value, we can be call it from Java.
  *
  * @see Generator
  *